Overly aggressive removal of overlays
I use a folding-mode which uses overlays, also preview-latex uses overlays and so does regular AUCTeX folding.  There are two places in `latex-preview-pane-update-p` where `remove-overlays` is called with no arguments.  This means that all those overlays are destroyed, when really only those from latex-preview-pane should be removed.  I believe the canonical way of doing this is to add a `name` attribute or something and key off of that, but I simply keyed off of font being `bad-face` which seems to work in practice (though bad-face should probably be renamed to have `lpp` in it somewhere).  I haven't tested the changes extensively, but my preview overlays stay.